Improving spell transferrence
Should be easier to extend now, as well as being less of a maintenance
nightmare. Also works better with IWDEE due to spwi801 being Monster
Summoning VI instead of a buff, and IWDIfication mod if SR is not
installed (especially in EET).

Also, did some minor scipr improvements which should improve IWDEE
compatibility.